Reduce chance of failsafe report failure
Opinion: "raygun4ruby encountered an exception processing your exception" is reported when the original exception fails to be reported. The goal of the failsafe report is to be as minimal as possible to increase the likelihood of successfully making its way to Raygun. With this in mind, it is my belief that the breadcrumbs store (and possibly other tertiary reporting features that I'm not aware of) should be cleared before reporting the failsafe message.

In this PR we simply discard extra state such as breadcrumbs when reporting the "raygun4ruby encountered an exception processing your exception", to reduce the chance of failsafe report failure.