Add size method to Breadcrumb
This calculates a rough size of the Breadcrumb, message length + 100.
Assuming all the other fields roughly come out to 100 characters on
average. The primary failing of this is that the metadata could be much
larger, but calculating the size of the metadata could be costly.

This handles the usecase needed by Vydia